feat(http): front controller + router + middleware (batch offset=0)
