+++ classes/Http/Handlers/PublicSite/ReputationHandler.php
@@ -1,0 +1,18 @@
+<?php
+declare(strict_types=1);
+
+namespace PU239\Http\Handlers\PublicSite;
+
+final class ReputationHandler
+{
+    public function handle(array $meta = []): mixed
+    {
+        if (!defined('PU239_ROUTED')) {
+            define('PU239_ROUTED', true);
+        }
+
+        require \dirname(__DIR__, 4) . '/public/reputation.php';
+
+        return null;
+    }
+}
