+++ classes/Security/PasswordHasher.php
@@ -45,4 +45,3 @@ final class PasswordHasher
     }
 }
 
-// >>>>>> PU239:pwdlight-helper-1
