+++ public/invite.php
@@ -1,4 +1,10 @@
 <?php
 declare(strict_types=1);
 
+if (!defined('PU239_ROUTED')) {
+    require_once __DIR__ . '/index.php';
+
+    return;
+}
+
 throw new RuntimeException('Stubbed: missing SQL; see tools/rehydrate_v3_manifest.csv');
