+++ staffpanel/index.php
@@ -1,6 +1,12 @@
 <?php
 declare(strict_types=1);
 
+if (!defined('PU239_ROUTED')) {
+    require_once __DIR__ . '/../public/index.php';
+
+    return;
+}
+
 require_once __DIR__ . '/../bootstrap_web.php';
 
 use Monolog\Logger;
